norm ok

+++ lib/ft_strnstr.c
@@ -6,13 +6,13 @@
 /*   By: gbethani <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/17 20:08:57 by gbethani          #+#    #+#             */
-/*   Updated: 2021/04/17 20:09:28 by gbethani         ###   ########.fr       */
+/*   Updated: 2021/04/18 16:10:20 by gbethani         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strnstr (const char *big, const char *little, size_t len)
+char	*ft_strnstr(const char *big, const char *little, size_t len)
 {
 	int	i;
 	int	j;
@@ -20,12 +20,12 @@ char	*ft_strnstr (const char *big, const char *little, size_t len)
 	if (little[0] == '\0')
 		return ((char *)big);
 	i = 0;
-	while (big[i] && i < len)
+	while (big[i] && i < (int)len)
 	{
 		j = 0;
 		if (big[i] == little[j])
 		{
-			while (i + j < len && big[i + j] == little[j])
+			while (i + j < (int)len && big[i + j] == little[j])
 			{
 				j++;
 				if (!little[j])
